Split component and doc example
